+++ CHANGELOG.md
@@ -44,6 +44,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `.jj/workspace-config.toml`. `jj config` subcommands which took layer options like
   `--repo` now also support `--workspace`.
 
+* Add support for `--when.workspaces` config scopes.
+
 * `jj bisect run` accepts the command and arguments to pass to the command
   directly as positional arguments, such as
   `jj bisect --range=..main -- cargo check --all-targets`.